# types/basetypes: Reduce memory allocations of (ObjectValue).ToTerraformValue()